Security: fix logging in to user ID 1 after signing up
Previously, :returning wasn't being parsed, and we were using execute
instead of query, so honeysql was returning the number of rows inserted,
not the generated :id.

This commit adds support for Postgres's RETURNING clause and uses the
returned :id in the session for company and dev members. Not visible
here is resetting the session cookie key to recycle all existing
sessions.